refactor: organize user, category and campaign routes
