+++ backend/routes/UsuariosRouter.js
@@ -1,0 +1,31 @@
+const express = require('express');
+const {
+  createUsuario,
+  loginUsuario,
+  asignarPuntos,
+  getPerfil,
+  getUsuarios,
+  getUsuario,
+  updateUsuario,
+  deleteUsuario,
+} = require('../controllers/UsuarioController');
+const authMiddleware = require('../middleware/authMiddleware');
+const adminMiddleware = require('../middleware/adminMiddleware');
+
+const router = express.Router();
+
+// Registro y login
+router.post('/usuarios', createUsuario);
+router.post('/auth/login', loginUsuario);
+
+// Perfil del usuario logueado
+router.get('/api/perfil', authMiddleware, getPerfil);
+
+// Acciones de administrador
+router.post('/api/admin/usuarios/:id/puntos', adminMiddleware, asignarPuntos);
+router.get('/api/admin/usuarios', adminMiddleware, getUsuarios);
+router.get('/api/admin/usuarios/:id', adminMiddleware, getUsuario);
+router.put('/api/admin/usuarios/:id', adminMiddleware, updateUsuario);
+router.delete('/api/admin/usuarios/:id', adminMiddleware, deleteUsuario);
+
+module.exports = router;
